+++ marple/src/dd/kms/marple/api/actions/InspectionAction.java
@@ -2,7 +2,6 @@ package dd.kms.marple.api.actions;
 
 public interface InspectionAction
 {
-	boolean isDefaultAction();
 	String getName();
 	String getDescription();
 	boolean isEnabled();