+++ marple/src/dd/kms/marple/impl/actions/AddVariableAction.java
@@ -24,11 +24,6 @@ public class AddVariableAction implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return false;
-	}
-
-	@Override
 	public String getName() {
 		return "Add to variables";
 	}