+++ marple/src/dd/kms/marple/impl/actions/CopyStringRepresentationAction.java
@@ -17,11 +17,6 @@ public class CopyStringRepresentationAction implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return false;
-	}
-
-	@Override
 	public String getName() {
 		return "Copy string representation";
 	}