+++ marple/src/dd/kms/marple/impl/actions/HighlightComponentAction.java
@@ -14,11 +14,6 @@ public class HighlightComponentAction implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return false;
-	}
-
-	@Override
 	public String getName() {
 		return "Highlight";
 	}