+++ marple/src/dd/kms/marple/impl/actions/HistoryForwardAction.java
@@ -21,11 +21,6 @@ public class HistoryForwardAction<T> implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return false;
-	}
-
-	@Override
 	public String getName() {
 		return "Forward";
 	}