+++ marple/src/dd/kms/marple/impl/actions/InspectObjectAction.java
@@ -16,11 +16,6 @@ public class InspectObjectAction implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return true;
-	}
-
-	@Override
 	public String getName() {
 		return "Inspect";
 	}