+++ marple/src/dd/kms/marple/impl/actions/InvokeMethodAction.java
@@ -20,11 +20,6 @@ public class InvokeMethodAction implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return true;
-	}
-
-	@Override
 	public String getName() {
 		return "Invoke";
 	}