+++ marple/src/dd/kms/marple/impl/gui/actionprovidertree/inspectiontree/ChangeNodeViewAction.java
@@ -19,11 +19,6 @@ class ChangeNodeViewAction implements InspectionAction
 	}
 
 	@Override
-	public boolean isDefaultAction() {
-		return false;
-	}
-
-	@Override
 	public String getName() {
 		return viewOption.getText();
 	}