+++ marple/src/dd/kms/marple/impl/gui/inspector/views/methodview/MethodViewUtils.java
@@ -27,15 +27,18 @@ class MethodViewUtils
 	ActionProvider getMethodActionProvider(Method method) {
 		ImmutableList.Builder<InspectionAction> actionsBuilder = ImmutableList.builder();
 		int numParameters = method.getParameterCount();
+		InspectionAction defaultAction = null;
 		if (numParameters == 0) {
-			actionsBuilder.add(createInvokeMethodAction(method));
+			InspectionAction invokeMethodAction = createInvokeMethodAction(method);
+			defaultAction = invokeMethodAction;
+			actionsBuilder.add(invokeMethodAction);
 		}
 		String methodName = method.getName();
 		String parameterPlaceholder = IntStream.range(0, numParameters).mapToObj(i -> "").collect(Collectors.joining(", "));
 		String expression = methodName + "(" + parameterPlaceholder + ")";
 		int caretPosition = numParameters == 0 ? expression.length() : methodName.length() + 1;
 		actionsBuilder.add(context.createEvaluateExpressionAction(expression, caretPosition, object));
-		return ActionProvider.of(methodName, actionsBuilder.build(), true);
+		return ActionProvider.of(methodName, actionsBuilder.build(), defaultAction);
 	}
 
 	private InspectionAction createInvokeMethodAction(Method method) {